move close_fd and dup2_fd to newfile
