+++ include/redirect.h
@@ -37,4 +37,8 @@ void delete_tmpfile(void);
 //heredoc_expansion.c
 char *heredoc_expantion(char *line, t_context *context);
 
+//redirect_utils.c
+void close_fd(int fd, t_context *context);
+void dup2_fd(int old_fd, int new_fd, t_context *context);
+
 #endif
